requirements.txt
torch
torchvision
pillow
matplotlib
tqdm
torchmetrics[image]
numpy
